refactor(shannon): split shannon_cmd into a directory module
2234 lines in one file, against the workspace's 800-line cap. Split by
concern, every file now well under it:

  args     254   subcommand enum and argument structs
  scoring  388   forward pass, per-token bits, summary
  arith    379   arithmetic coder, bit IO, `.shannon` container
  verify   349   cross-engine comparison and the reference scorers
  vindex   331   encode/decode straight from a quantised index
  layers   315   layer/slot/repeat diagnostics
  mod      239   shared constants and the command dispatcher

Strictly mechanical: items were moved whole by name with brace matching,
not carved by line range — a line-range carve is what went wrong on the
shader_bench split and got reverted. No item was edited beyond the
visibility each move forced (`pub(super)` on items and struct fields that
are now referenced across the boundary), and the two paths that leave the
module keep their original visibility via targeted re-exports rather than
a widened glob.

Behaviour-preserving, checked rather than asserted: SmolLM2-135M over the
1 KB corpus scores 273 tokens / 3.036 bits per token / 828.9 total bits
before and after, and 694 larql-cli tests pass. fmt and
`clippy --all-targets -D warnings` clean.

No functional change — the packed-MoE work that motivated touching this
file lands separately on top, so the substantive history stays bisectable.
